+++ src/test/java/de/haupz/basicode/IntTest.java
@@ -6,14 +6,14 @@ public class IntTest extends ExpressionTest {
 
     @Test
     public void testInteger() {
-        testExpression("INT(1)", 1, Integer.class);
-        testExpression("INT(-1)", -1, Integer.class);
+        testExpression("INT(1)", 1.0, Double.class);
+        testExpression("INT(-1)", -1.0, Double.class);
     }
 
     @Test
     public void testDouble() {
-        testExpression("INT(1.1)", 1, Integer.class);
-        testExpression("INT(-1.1)", -2, Integer.class);
+        testExpression("INT(1.1)", 1.0, Double.class);
+        testExpression("INT(-1.1)", -2.0, Double.class);
     }
 
     @Test
